users/pyproject.toml
[tool.poetry]
name = "users"
version = "0.0.1"
description = ""
authors = ["Iskander Shamsutdinov <wensietyt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
alembic = "^1.13.3"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
uvicorn = "^0.31.0"
pylint = "^3.3.1"
mypy = "^1.11.2"
flake8 = "^7.1.1"
psycopg2-binary = "^2.9.9"
isort = "^5.13.2"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
